+++ src/types.rs
@@ -6,7 +6,7 @@ use std::{
 #[allow(dead_code)]
 #[derive(Clone, Debug, PartialEq, PartialOrd)]
 pub enum Numeric {
-    Int(i128), // dependiendo del build, tendras: 32 o 64 bits de numero entero
+    Int(i64), // dependiendo del build, tendras: 32 o 64 bits de numero entero
     Float(f64),
 }
